+++ browser/src/Services/IconThemes/Icons.ts
@@ -11,8 +11,8 @@ import { Event, IEvent } from "oni-types"
 
 import { PluginManager } from "./../../Plugins/PluginManager"
 
-// import { IConfigurationValues } from "./Configuration"
-// import { getThemeManagerInstance, ThemeManager } from "./Themes"
+import { StyleWriter } from "./StyleWriter"
+import { PluginIconThemeLoader } from "./IconThemeLoader"
 
 export interface IIconFontSource {
     path: string
@@ -57,10 +57,8 @@ export interface IIconTheme {
     fileExtensions: FileDefinitions
     fileNames: FileNames
     languageIds: Language
+    light?: IIconTheme
 }
-
-import * as fs from "fs"
-import { IIconThemeContribution } from "./../../Plugins/Api/Capabilities"
 
 export class Icons {
 
@@ -129,45 +127,14 @@ export class Icons {
 
     public async applyIconTheme(themeName: string): Promise<void> {
 
-        const plugins = this._pluginManager.plugins
+        const iconThemeLoader = new PluginIconThemeLoader(this._pluginManager)
 
-        const pluginsWithThemes = plugins.filter((p) => {
-            return p.metadata && p.metadata.contributes && p.metadata.contributes.iconThemes
-        })
-
-        const allIconThemes = pluginsWithThemes.reduce((previous: IIconThemeContribution[], current) => {
-            const iconThemes = current.metadata.contributes.iconThemes
-            return [
-                ...previous,
-                ...iconThemes
-            ]
-        }, [] as IIconThemeContribution[])
-
-        const matchingIconTheme = allIconThemes.find((t) => t.id === themeName)
-
-        if (!matchingIconTheme || !matchingIconTheme.path) {
-            return null
-        }
-
-        const contents = await new Promise<string>((resolve, reject) => {
-            fs.readFile(matchingIconTheme.path, "utf8", (err, data: string) => {
-                if (err) {
-                    reject(err)
-                    return
-                }
-
-                resolve(data)
-            })
-        })
-
-        // From stackoverflow: 
-        // https://stackoverflow.com/questions/11355147/font-face-changing-via-javascript
-        this._activeIconTheme = JSON.parse(contents) as IIconTheme
+        this._activeIconTheme = await iconThemeLoader.loadIconTheme(themeName)
 
         var newStyle = document.createElement("style")
-
         const styleWriter = new StyleWriter("oni-icon")
 
+        // TODO: Path
         styleWriter.writeFontFace("seti","C:/oni/extensions/theme-icons-seti/icons/seti.woff", "woff")
 
         const iconDefinitions = this._activeIconTheme.iconDefinitions
@@ -183,67 +150,5 @@ export class Icons {
         document.head.appendChild(newStyle)
 
         this._onIconThemeChangedEvent.dispatch()
-    }
-}
-
-import * as os from "os"
-
-export class StyleWriter {
-    private _style: string = ""
-
-    public get style(): string {
-        return this._style
-    }
-
-    constructor(
-        private _primaryClassName: string
-    ) {}
-
-    public writeFontFace(fontFamily: string, sourceUrl: string, format: string): void {
-        // Inspired by:
-        // https://stackoverflow.com/questions/11355147/font-face-changing-via-javascript
-        const fontFaceBlock = [
-            "@font-face {",
-            `   font-family: ${fontFamily};`,
-            `   src: url('${sourceUrl}') format('${format}');`,
-            "}",
-        ]
-
-        this._append(fontFaceBlock)
-
-        const primaryClassBlock = [
-            ".fa." + this._primaryClassName + " {",
-                "font-family: " + fontFamily + ";",
-            "}",
-        ]
-
-        this._append(primaryClassBlock)
-    }
-
-    public writeIcon(iconName: string, fontColor: string, fontCharacter: string): void {
-        const iconClass = this._primaryClassName + "-" + iconName
-        const selector = ".fa." + this._primaryClassName + "." + iconClass
-
-        if (fontColor) {
-            const primaryClassBlock = [
-                selector + " {",
-                    "color: " + fontColor + ";",
-                "}"
-            ]
-            this._append(primaryClassBlock)
-        }
-
-        const pseudoElementBlock = [
-            selector + ":before {",
-            `   content: '${fontCharacter}';`,
-            "}"
-        ]
-        this._append(pseudoElementBlock)
-    }
-
-
-
-    private _append(str: string[]): void {
-        this._style += str.join(os.EOL) + os.EOL
     }
 }
